test: cover bot-api verbose secret key redaction
Extract shouldRedactKey/redactSecrets and sanitizeErrorText into shared
leaf modules so verbose Next→bot logs cannot regress and leak tokens.
